+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 # compose-auto-preview Changelog
 
+## [2.0.0]
+
+### UI Changes
+
+Now, there is only one "Generate Compose Preview", that opens a popup to choose what type of Preview should be generated.
+
+### New Preview Types
+
+* Added support for Previews with System UI.
+* Added support for dark mode, for all preview types
+
 ## [1.2.0]
 - Remove deprecated functions.